LayoutTests/inspector/styles/stylesheet-tracking-expected.txt

Tests that the styleSheetAdded and styleSheetRemoved events are reported into the frontend. Bug 105828.

Text
2 headers known:
0:
    URL=.../inspector/styles/resources/stylesheet-tracking.css
    origin=regular
1:
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=regular
Adding stylesheet...
Stylesheet added:
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=regular
Removing stylesheet...
Stylesheet removed:
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=regular
Adding rule...
Stylesheet added:
    URL=.../inspector/styles/stylesheet-tracking.html
    origin=inspector
Rule added
